Moving TF common files to separated directory
