+++ src/python/modules/Tensorflow/TensorflowGMM.py
@@ -3,10 +3,10 @@ from __future__ import absolute_import, division, print_function, \
                        
 import tensorflow as tf
 
-from modules.Tensorflow.utils import to_tf_tensor, flatten
+from modules.TensorflowCommon.utils import to_tf_tensor, flatten
 from shared.ITest import ITest
 from shared.GMMData import GMMInput, GMMOutput
-from modules.Tensorflow.gmm_objective import gmm_objective
+from modules.TensorflowCommon.gmm_objective import gmm_objective
