+++ src/python/modules/TensorflowCommon/hand_objective.py
@@ -1,5 +1,5 @@
 import tensorflow as tf
-from modules.Tensorflow.utils import shape
+from modules.TensorflowCommon.utils import shape
 
 
 def to_pose_params(theta, nbones):
